+++ pyogrio/_compat.py
@@ -24,13 +24,16 @@ except ImportError:
     pandas = None
 
 
-HAS_ARROW_API = __gdal_version__ >= (3, 6, 0) and pyarrow is not None
+HAS_ARROW_API = __gdal_version__ >= (3, 6, 0)
+HAS_PYARROW = pyarrow is not None
 
 HAS_GEOPANDAS = geopandas is not None
 
 PANDAS_GE_15 = pandas is not None and Version(pandas.__version__) >= Version("1.5.0")
 PANDAS_GE_20 = pandas is not None and Version(pandas.__version__) >= Version("2.0.0")
 PANDAS_GE_22 = pandas is not None and Version(pandas.__version__) >= Version("2.2.0")
+
+GDAL_GE_38 = __gdal_version__ >= (3, 8, 0)
 
 HAS_GDAL_GEOS = __gdal_geos_version__ is not None